+++ BabylonNativeCs/BabylonNativeCsLibrary/PostProcess/babylon.anaglyphPostProcess.cs
@@ -7,7 +7,7 @@ namespace BABYLON
 {
     public partial class AnaglyphPostProcess : PostProcess
     {
-        public AnaglyphPostProcess(string name, double ratio, Camera camera, double samplingMode = 0.0, Engine engine = null, bool reusable = false)
+        public AnaglyphPostProcess(string name, double ratio, Camera camera, int samplingMode = 0, Engine engine = null, bool reusable = false)
             : base(name, "anaglyph", null, new Array<string>("leftSampler"), ratio, camera, samplingMode, engine, reusable)
         {
         }
