+++ BabylonNativeCs/BabylonNativeCsLibrary/PostProcess/babylon.blackAndWhitePostProcess.cs
@@ -3,8 +3,11 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using Web;
-namespace BABYLON {
-    public partial class BlackAndWhitePostProcess: PostProcess {
-        public BlackAndWhitePostProcess(string name, double ratio, Camera camera, double samplingMode = 0.0, Engine engine = null, bool reusable = false): base(name, "blackAndWhite", null, null, ratio, camera, samplingMode, engine, reusable) {}
+namespace BABYLON
+{
+    public partial class BlackAndWhitePostProcess : PostProcess
+    {
+        public BlackAndWhitePostProcess(string name, double ratio, Camera camera, int samplingMode = 0, Engine engine = null, bool reusable = false) 
+            : base(name, "blackAndWhite", null, null, ratio, camera, samplingMode, engine, reusable) { }
     }
 }
